[C8] Split dressup.go per dress-up feature (#2233)
model/feature/dressup.go was 524 LOC, over the 500-line ceiling. Split the
dress-up feature definitions per feature, keeping the DressUpFeatures adder
collection and the shared anchor capture in dressup.go; each file is now
under the ceiling:

- dressup.go (203): the DressUpFeatures adder collection (Add*/add* methods)
  and the shared resolveFacesThenDefer / runningBody helpers.
- dressup_fillet.go (108): the fillet definition + feature.
- dressup_chamfer.go (73): the chamfer definition + feature.
- dressup_shell.go (51): the shell definition + feature.
- dressup_draft.go (39): the face-draft definition + feature.
- dressup_thread.go (120): the thread definition + feature.

Behaviour-preserving: split programmatically by receiver-qualified top-level
declaration, then verified every non-comment code line is byte-identical to
the original (only per-file import redistribution + gofmt one-liner alignment
differ, the latter restored); imports resolved per file with goimports.

The .golangci.yml funlen/gocyclo exclude-rule for model/feature/dressup.go is
broadened to dressup[_a-z0-9]*.go so the split files inherit the SAME
pre-existing exclusion (dress-up construction is intentionally linear) — the
byte-identical thread Recompute was excluded before and stays excluded. This
also pre-covers the loft_skin/sweep_loft/swept split tasks. Feature corpus +
archguard pass; lint 0. Net delta of recognizers / tolerance constants /
fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2233.

(cherry picked from commit 92ed0bb7526cfba4ba7f3740ab6cc4cd4f6cc543)
