+++ model/feature/dressup.go
@@ -8,7 +8,6 @@ import (
 
 	"oblikovati.org/api/types"
 	"oblikovati.org/kernel/geom"
-	"oblikovati.org/kernel/ops"
 	"oblikovati.org/kernel/topo"
 	"oblikovati.org/math"
 )
@@ -21,326 +20,6 @@ import (
 // (parametric-cad §7, ADR-0010). The rounding/cut geometry itself is kernel phase B
 // (rolling-ball fillets), so once inputs resolve these features report
 // [ErrDeferred] (→ health.Warning) and pass the body through unchanged.
-
-// FilletEdgeSet is one edge set of a fillet definition (the reference's
-// FilletConstantRadiusEdgeSet / FilletVariableRadiusEdgeSet): a constant Radius over the
-// whole set, or — when Radius is nil — a variable StartRadius→EndRadius over a single edge
-// (radius runs linearly from the edge's start vertex to its end vertex). RadiusPoints are
-// optional intermediate (position, radius) stops along that variable edge (#695), each at a
-// fraction 0<T<1 of the edge, strictly increasing in T — the reference's FilletRadius set points.
-type FilletEdgeSet struct {
-	EdgeKeys     [][]byte
-	Radius       func() float64
-	StartRadius  func() float64
-	EndRadius    func() float64
-	RadiusPoints []FilletRadiusPoint
-}
-
-// FilletRadiusPoint is one intermediate radius stop on a variable fillet edge: T is the fraction
-// along the edge (start vertex = 0, end vertex = 1), Radius the rolling-ball radius there.
-type FilletRadiusPoint struct {
-	T      float64
-	Radius func() float64
-}
-
-// variable reports whether the set carries a start→end radius instead of a constant one.
-func (s FilletEdgeSet) variable() bool { return s.Radius == nil }
-
-// FilletCornerType aliases the public corner-treatment discriminator (ADR-0018).
-type FilletCornerType = types.FilletCornerType
-
-// FilletCrossSection aliases the public blend cross-section shape (M36-F08, ADR-0018): arc (G1,
-// default), G2 (curvature-continuous), or conic (rho-controlled).
-type FilletCrossSection = types.FilletCrossSection
-
-// Fillet cross-section shapes (aliases of the canonical api/types values).
-const (
-	FilletArc   = types.FilletSectionArc
-	FilletG2    = types.FilletSectionG2
-	FilletConic = types.FilletSectionConic
-)
-
-// FilletDefinition rounds selected edges. EdgeKeys+Radius is the original single
-// constant-radius form; EdgeSets (when non-empty) takes precedence and carries any mix of
-// constant and variable sets (#323). CornerType selects how a vertex where two filleted edges
-// meet (third edge sharp) is treated — miter crease (default/zero), or a full round.
-type FilletDefinition struct {
-	EdgeKeys        [][]byte
-	Radius          func() float64
-	EdgeSets        []FilletEdgeSet
-	CornerType      FilletCornerType
-	ConcaveStrategy types.FilletConcaveStrategy // concave edges: outward fill (zero/default) or inward recess
-	// CrossSection selects the blend cross-section shape (M36-F08): the default arc (G1), a
-	// curvature-continuous G2, or a rho-controlled conic. Rho sets a conic's fullness (0<ρ<1,
-	// 0.5=parabola). Non-arc sections build via the swept ruling band (no analytic cylinder).
-	CrossSection FilletCrossSection
-	Rho          float64
-	// GeomEdges are edges selected by a serialized GEOMETRIC descriptor rather than an
-	// Oblikovati lineage key — the path an external author (the NX exporter, M8/ADR-0040)
-	// uses, since it cannot synthesize lineage. They bind to the running body's edges at
-	// recompute (see bindGeomEdges) and fold into the edge list; absent for a normal
-	// Oblikovati-authored fillet.
-	GeomEdges []topo.GeometricEdgeRef
-	// EdgeAnchors maps an EdgeKeys entry (raw reference key, as a string) to the edge's
-	// midpoint captured when the user picked it. It feeds the GEOMETRIC recovery tier
-	// (ADR-0043 P6b): when a lost key's parent has several surviving siblings, the anchor
-	// disambiguates by nearness. Absent for an older recipe or an edit-mode retained key —
-	// such a reference degrades to exact/ancestral recovery only.
-	EdgeAnchors map[string]math.Point3
-}
-
-// FilletType reports the definition's discriminator: always an edge fillet for now (the
-// reference's face and full-round fillets are follow-ups tracked on #323).
-func (d *FilletDefinition) FilletType() types.FilletType { return types.EdgeFillet }
-
-// FilletFeature is an edge fillet over one or more constant/variable radius edge sets.
-type FilletFeature struct{ def *FilletDefinition }
-
-// Definition returns the fillet recipe.
-func (f *FilletFeature) Definition() *FilletDefinition { return f.def }
-
-// Kind implements [Feature].
-func (f *FilletFeature) Kind() string { return "fillet" }
-
-// Recompute rounds the picked convex edges on the running body with a real rolling-ball
-// blend (cylinder faces; planar ruling strips for variable sets). See fillet.go.
-func (f *FilletFeature) Recompute(in Input) (Output, error) {
-	prof := blendProfile{cross: f.def.CrossSection, rho: f.def.Rho}
-	if len(f.def.EdgeSets) > 0 {
-		return filletBodySets(in, f.def.EdgeSets, f.def.CornerType, f.def.ConcaveStrategy, prof, "fillet")
-	}
-	keys, err := bindGeomEdges(in, f.def.EdgeKeys, f.def.GeomEdges, "fillet")
-	if err != nil {
-		return Output{}, err
-	}
-	return filletBody(in, keys, callOrZero(f.def.Radius), f.def.CornerType, f.def.ConcaveStrategy, prof, "fillet", f.def.EdgeAnchors)
-}
-
-// ChamferType aliases the public chamfer-mode discriminator (ADR-0018).
-type ChamferType = types.ChamferType
-
-// ChamferConcaveStrategy aliases the public concave-edge strategy discriminator (ADR-0018).
-type ChamferConcaveStrategy = types.ChamferConcaveStrategy
-
-// ChamferDefinition bevels selected edges. Type (M20-F03) selects the setback mode: equal
-// Distance on both faces (the default / zero value), two independent distances (Distance +
-// Distance2, asymmetric), or a Distance plus the chamfer-face Angle. FlatCorners blends a
-// vertex where three selected edges meet into a flat triangular face — only for the
-// equal-distance mode; an asymmetric chamfer leaves the corner planes to meet at a point.
-// ConcaveStrategy applies only to CONCAVE (internal) edges: fill the inside corner with material
-// (outward, the zero-value default) or cut a recessed relief groove (inward).
-type ChamferDefinition struct {
-	EdgeKeys        [][]byte
-	Distance        func() float64
-	Distance2       func() float64 // twoDistances: setback on the second face
-	Angle           func() float64 // distanceAndAngle: chamfer-face angle (radians)
-	Type            ChamferType    // zero value ⇒ equal-distance
-	FlatCorners     bool
-	ConcaveStrategy ChamferConcaveStrategy  // zero value ⇒ outward (fill the inside corner)
-	GeomEdges       []topo.GeometricEdgeRef // externally-authored edges by geometric descriptor (see FilletDefinition.GeomEdges)
-	// EdgeAnchors maps an EdgeKeys entry to its mint-time midpoint for the geometric recovery
-	// tier (ADR-0043 P6b); see FilletDefinition.EdgeAnchors.
-	EdgeAnchors map[string]math.Point3
-	// ReferenceFace is the face Distance is measured on for the asymmetric modes (#1888). Empty
-	// leaves the assignment to the edge's own face order, which is a topology artefact — on
-	// mirrored geometry that can put the larger setback on the wrong face. See orderedSetbacks.
-	ReferenceFace []byte
-	// PartialStart and PartialLength bevel only a SPAN of each edge, measured from its start vertex
-	// (#1888). PartialLength 0 ⇒ the whole edge. See wedgeSpan.
-	PartialStart  func() float64
-	PartialLength func() float64
-}
-
-// ChamferFeature bevels selected edges (equal-distance, two-distance, or distance-and-angle).
-type ChamferFeature struct {
-	def      *ChamferDefinition
-	featName string
-}
-
-func (c *ChamferFeature) Definition() *ChamferDefinition { return c.def }
-func (c *ChamferFeature) Kind() string                   { return "chamfer" }
-
-// Recompute bevels each selected (convex) edge by cutting a wedge tool along it via the
-// boolean; the two setbacks come from the mode (see chamferSetbacks). Flat-corner blending
-// applies to every mode — the blend is built from the per-face setbacks, so an asymmetric
-// (two-distance / distance-angle) corner blends just like a symmetric one. See chamfer.go.
-func (c *ChamferFeature) Recompute(in Input) (Output, error) {
-	d1, d2, err := chamferSetbacks(c.def)
-	if err != nil {
-		return Output{}, err
-	}
-	keys, err := bindGeomEdges(in, c.def.EdgeKeys, c.def.GeomEdges, "chamfer")
-	if err != nil {
-		return Output{}, err
-	}
-	return chamferEdges(in, keys, d1, d2, featOr(c.featName, "chamfer"), c.def.FlatCorners,
-		c.def.ConcaveStrategy, c.def.runOf(), c.def.EdgeAnchors)
-}
-
-// ShellDefinition hollows a body, removing the selected faces, to a wall thickness.
-type ShellDefinition struct {
-	RemovedFaceKeys [][]byte
-	Thickness       func() float64
-	GeomFaces       []topo.GeometricFaceRef // externally-authored removed faces by geometric descriptor (ADR-0040)
-	Direction       ops.ShellDirection      // which side the wall grows onto (default ShellInside) — #1864
-	// FaceThicknesses give named RETAINED faces their own wall thickness (Inventor's
-	// SetFaceThickness, #1864) — a thickened boss wall or a thin window in an otherwise uniform
-	// shell. Each is a closure so a parameter drives it, exactly like the default Thickness.
-	FaceThicknesses []ShellFaceThickness
-}
-
-// ShellFaceThickness overrides the wall thickness on one retained face of a shell (#1864).
-type ShellFaceThickness struct {
-	FaceKey   []byte
-	Thickness func() float64
-}
-
-// ShellFeature hollows a solid.
-type ShellFeature struct {
-	def      *ShellDefinition
-	featName string
-}
-
-func (s *ShellFeature) Definition() *ShellDefinition { return s.def }
-func (s *ShellFeature) Kind() string                 { return "shell" }
-
-// Recompute hollows the running body to the wall thickness, opening the removed faces. See
-// shell.go.
-func (s *ShellFeature) Recompute(in Input) (Output, error) {
-	keys, err := bindGeomFaces(in, s.def.RemovedFaceKeys, s.def.GeomFaces, featOr(s.featName, "shell"))
-	if err != nil {
-		return Output{}, err
-	}
-	return shellBody(in, keys, callOrZero(s.def.Thickness), s.def.Direction,
-		evalFaceThicknesses(s.def.FaceThicknesses), featOr(s.featName, "shell"))
-}
-
-// FaceDraftDefinition tapers selected faces by an angle about a pull direction. Neutral, when set, is
-// the fixed parting plane each face pivots on (#1801); nil ⇒ the implicit lowest-vertex hinge.
-type FaceDraftDefinition struct {
-	FaceKeys  [][]byte
-	PullDir   math.Vector3
-	Neutral   *geom.Plane
-	Angle     func() float64
-	GeomFaces []topo.GeometricFaceRef // externally-authored drafted faces by geometric descriptor (ADR-0040)
-}
-
-// FaceDraftFeature applies draft to faces.
-type FaceDraftFeature struct{ def *FaceDraftDefinition }
-
-func (d *FaceDraftFeature) Definition() *FaceDraftDefinition { return d.def }
-func (d *FaceDraftFeature) Kind() string                     { return "draft" }
-
-// Recompute tapers the picked faces about the pull direction by the angle (see draft.go).
-func (d *FaceDraftFeature) Recompute(in Input) (Output, error) {
-	keys, err := bindGeomFaces(in, d.def.FaceKeys, d.def.GeomFaces, "draft")
-	if err != nil {
-		return Output{}, err
-	}
-	return draftBody(in, keys, d.def.PullDir, d.def.Neutral, callOrZero(d.def.Angle), "draft")
-}
-
-// ThreadDefinition applies thread data to a cylindrical face. Cut=false is a cosmetic thread
-// (data + display, solid unchanged); Cut=true models a real thread (a helical groove cut).
-// Class, Tapered, and ModelDiameter are the #325 parity fields: the tolerance class recorded
-// on the spec, the pipe-thread flag (the reference's TaperedThreadInfo split — data-only, a
-// cut tapered thread needs a conical face and errors), and which thread diameter the modeled
-// cylindrical face represents (zero value = major, the common case).
-type ThreadDefinition struct {
-	FaceKey       []byte
-	Designation   string
-	Cut           bool
-	Class         string
-	Tapered       bool
-	ModelDiameter types.ModelDiameterFromThread
-	// Offset and Length limit the thread to a sub-span of the face along its axis (Inventor's
-	// ThreadOffset / ThreadDepth): the thread runs from the face's min axial extent + Offset for
-	// Length (cm). A nil Offset ⇒ 0; a nil or zero Length ⇒ the full face (Inventor's FullDepth).
-	// A double-ended stud threads only its two ends by giving each thread its own Offset+Length.
-	Offset func() float64
-	Length func() float64
-	// LeftHanded reverses the thread's sense (#1892). Named for the LEFT hand, not Inventor's
-	// RightHanded, for the same reason as HoleTap.LeftHanded: a definition built as a literal
-	// would otherwise default every thread to left-handed. A "-LH" designation says the same
-	// thing; either one alone makes the thread left-handed.
-	LeftHanded bool
-	// FaceAnchors maps FaceKey to its mint-time centroid for the geometric recovery tier
-	// (ADR-0043 P6 / #1579); see FilletDefinition.EdgeAnchors.
-	FaceAnchors map[string]math.Point3
-}
-
-// ThreadFeature tags a cylindrical face with a cosmetic thread (Inventor's ThreadFeature): it
-// records the resolved thread data and leaves the solid unchanged. Cut-thread geometry (a real
-// helical groove) is a separate modeled feature; the cosmetic thread is the data + display.
-type ThreadFeature struct {
-	def  *ThreadDefinition
-	spec *ThreadSpec // resolved on the last recompute (nil until then)
-}
-
-func (t *ThreadFeature) Definition() *ThreadDefinition { return t.def }
-func (t *ThreadFeature) Kind() string                  { return "thread" }
-
-// Spec returns the thread data resolved on the last recompute (nil if it never bound).
-func (t *ThreadFeature) Spec() *ThreadSpec { return t.spec }
-
-// Recompute parses the designation, binds the cylindrical face, records the thread spec, and
-// passes the (unchanged) solid through. A bad designation, a lost face, or a non-cylindrical
-// face makes the feature Sick — as does cutting a tapered (pipe) thread, which would need a
-// conical face the feature doesn't model yet.
-func (t *ThreadFeature) Recompute(in Input) (Output, error) {
-	spec, err := ParseThreadDesignation(t.def.Designation)
-	if err != nil {
-		return Output{}, err
-	}
-	spec.Class, spec.Tapered = t.def.Class, t.def.Tapered
-	// The designation may already carry the handedness as an "-LH" suffix, so the flag can only
-	// take handedness AWAY from right — the two spellings agree rather than fight (#1892).
-	spec.RightHanded = spec.RightHanded && !t.def.LeftHanded
-	spec.ModelDiameter = t.def.ModelDiameter
-	if spec.ModelDiameter == 0 {
-		spec.ModelDiameter = types.ThreadMajorDiameter
-	}
-	if t.def.Tapered && t.def.Cut {
-		return Output{}, fmt.Errorf("thread %q: a cut tapered (pipe) thread needs a conical face; model it cosmetic", t.def.Designation)
-	}
-	body, err := runningBody(in)
-	if err != nil {
-		return Output{}, err
-	}
-	face, mt, err := bindFace(body, t.def.FaceKey, anchorFor(t.def.FaceKey, t.def.FaceAnchors))
-	if err != nil {
-		return Output{}, fmt.Errorf("thread: %w", err)
-	}
-	cyl, ok := face.Geometry().(geom.Cylinder)
-	if !ok {
-		return Output{}, fmt.Errorf("thread %q: face is not cylindrical (%T)", t.def.Designation, face.Geometry())
-	}
-	vFaceMin, vFaceMax := axialExtent(face.RangeBox(), cyl)
-	vMin, vMax := resolveThreadSpan(vFaceMin, vFaceMax, t.def.Offset, t.def.Length)
-	spec.Internal = bodyHasMaterialOutside(body, cyl, (vMin+vMax)/2, (spec.MajorDiameter-spec.MinorDiameter)/2/10)
-	t.spec = &spec
-	if !t.def.Cut {
-		return Output{Bodies: in.Bodies, Heals: faceHeal(t.def.FaceKey, mt)}, nil // cosmetic: solid unchanged
-	}
-	// Modeled (cut) thread: retype the cylindrical face to a threaded surface — O(1), no
-	// boolean — so it tessellates and measures as real threaded geometry. The span honours the
-	// thread's offset/length so a partial cut thread grooves only its run.
-	threaded := geom.ThreadedCylinder{
-		Cylinder: cyl, Pitch: spec.Pitch / 10, Depth: (spec.MajorDiameter - spec.MinorDiameter) / 2 / 10,
-		Designation: t.def.Designation, Internal: spec.Internal, RightHanded: spec.RightHanded,
-		VMin: vMin, VMax: vMax,
-	}
-	out := make([]*topo.Body, len(in.Bodies))
-	copy(out, in.Bodies)
-	// Target the RESOLVED face's current key, not the stored one: a healed thread bound to a
-	// recovered sibling whose live key differs from t.def.FaceKey (ADR-0043 P6, mirrors edges).
-	threadedBody, err := ops.ReplaceFaceSurface(body, face.ReferenceKey(), threaded)
-	if err != nil {
-		return Output{}, err
-	}
-	out[len(out)-1] = threadedBody // runningBody is the last body
-	return Output{Bodies: out, Heals: faceHeal(t.def.FaceKey, mt)}, nil
-}
 
 // resolveFacesThenDefer resolves face keys against the running body and, if all bind, defers
 // the geometry (passthrough + ErrDeferred); a lost key is a Sick error. (Still used by the
